--- google_cloud_automlops/orchestration/enums.py
@@ -1,35 +1,0 @@
-# Copyright 2023 Google LLC. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-"""Sets global enums."""
-
-# pylint: disable=C0103
-# pylint: disable=line-too-long
-
-from enum import Enum
-
-class Orchestrator(Enum):
-    """Enum representing the available options for orchestration management."""
-
-    KFP = 'kfp'
-    # ARGO_WORKFLOWS = 'argo-workflows'   # roadmap item
-    # TFX = 'tfx'   # roadmap item
-    # AIRFLOW = 'airflow'   # roadmap item
-    # RAY = 'ray'   # roadmap item
-
-class PipelineJobSubmitter(Enum):
-    """Enum representing the available options for the Pipeline Job submission service."""
-
-    CLOUD_FUNCTIONS = 'cloud-functions'
-    CLOUD_RUN = 'cloud-run'
