+++ google_cloud_automlops/orchestration/kfp.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Creates a KFP pipeline subclass."""
+"""Creates a KFP component, pipeline, and services subclass."""
 
 # pylint: disable=anomalous-backslash-in-string
 # pylint: disable=C0103
@@ -21,7 +21,7 @@
 import json
 import re
 import textwrap
-from typing import Callable, Optional
+from typing import Callable, List, Optional
 
 try:
     from importlib.resources import files as import_files
@@ -29,7 +29,7 @@ except ImportError:
     # Try backported to PY<37 `importlib_resources`
     from importlib_resources import files as import_files
 
-from google_cloud_automlops.orchestration.Pipeline import Pipeline
+from google_cloud_automlops.orchestration.base import BaseComponent, BasePipeline, BaseServices
 from google_cloud_automlops.utils.utils import (
     execute_process,
     get_components_list,
@@ -39,6 +39,7 @@ from google_cloud_automlops.utils.utils import (
     render_jinja,
     write_and_chmod,
     write_file,
+    write_yaml_file
 )
 from google_cloud_automlops.utils.constants import (
     BASE_DIR,
@@ -56,11 +57,182 @@ from google_cloud_automlops.utils.constants import (
     GENERATED_RUN_ALL_SH_FILE,
     KFP_TEMPLATES_PATH,
     PINNED_KFP_VERSION,
-    PIPELINE_CACHE_FILE
+    PLACEHOLDER_IMAGE
 )
 
 
-class KFPPipeline(Pipeline):
+class KFPComponent(BaseComponent):
+    """Creates a KFP specific Component object for #TODO: add more
+
+    Args:
+        Component (object): Generic Component object.
+    """
+
+    def __init__(self,
+                 func: Optional[Callable] = None, 
+                 packages_to_install: Optional[List[str]] = None):
+        """Initiates a KFP Component object created out of a function holding
+        all necessary code.
+
+        Args:
+            func: The python function to create a component from. The function
+                should have type annotations for all its arguments, indicating how
+                it is intended to be used (e.g. as an input/output Artifact object,
+                a plain parameter, or a path to a file).
+            packages_to_install: A list of optional packages to install before
+                executing func. These will always be installed at component runtime.
+        """
+        super().__init__(func, packages_to_install)
+
+        # Update parameters and return types to reflect KFP data types
+        if self.parameters:
+            self.parameters = self._update_params(self.parameters)
+        if self.return_types:
+            self.return_types = self._update_params(self.return_types)
+
+        # Set packages to install and component spec attributes
+        self.packages_to_install_command = self._get_packages_to_install_command()
+        self.component_spec = self._create_component_spec()
+
+    def build(self):
+        """Constructs files for running and managing Kubeflow pipelines.
+        """
+        defaults = read_yaml_file(GENERATED_DEFAULTS_FILE)
+        self.artifact_repo_location = defaults['gcp']['artifact_repo_location']
+        self.artifact_repo_name = defaults['gcp']['artifact_repo_name']
+        self.project_id = defaults['gcp']['project_id']
+        self.naming_prefix = defaults['gcp']['naming_prefix']
+
+        # Set and create directory for components if it does not already exist
+        component_dir = BASE_DIR + 'components/' + self.component_spec['name']
+
+        # Build necessary folders
+        # TODO: make this only happen for the first component? or pull into automlops.py
+        make_dirs([
+            component_dir,
+            BASE_DIR + 'components/component_base/src/'])
+
+        # TODO: can this be removed?
+        kfp_spec_bool = self.component_spec['implementation']['container']['image'] != PLACEHOLDER_IMAGE
+
+        # Read in component specs
+        custom_code_contents = self.component_spec['implementation']['container']['command'][-1]
+        compspec_image = (
+                f'''{self.artifact_repo_location}-docker.pkg.dev/'''
+                f'''{self.project_id}/'''
+                f'''{self.artifact_repo_name}/'''
+                f'''{self.naming_prefix}/'''
+                f'''components/component_base:latest''')
+
+        # If using kfp, remove spaces in name and convert to lowercase
+        if kfp_spec_bool:
+            self.component_spec['name'] = self.component_spec['name'].replace(' ', '_').lower()
+
+        # Write task script to component base
+        write_file(
+            filepath=BASE_DIR + 'components/component_base/src/' + self.component_spec['name'] + '.py',
+            text=render_jinja(
+                template_path=import_files(KFP_TEMPLATES_PATH + '.components.component_base.src') / 'task.py.j2',
+                generated_license=GENERATED_LICENSE,
+                kfp_spec_bool=kfp_spec_bool,
+                custom_code_contents=custom_code_contents),
+            mode='w')
+
+        # Update component_spec to include correct image and startup command
+        self.component_spec['implementation']['container']['image'] = compspec_image
+        self.component_spec['implementation']['container']['command'] = [
+            'python3',
+            f'''/pipelines/component/src/{self.component_spec['name']+'.py'}''']
+
+        # Write license and component spec to the appropriate component.yaml file
+        comp_yaml_path = component_dir + '/component.yaml'
+        write_file(
+            filepath=comp_yaml_path,
+            text=GENERATED_LICENSE,
+            mode='w')
+        write_yaml_file(
+            filepath=comp_yaml_path,
+            contents=self.component_spec,
+            mode='a')
+
+    def _get_packages_to_install_command(self):
+        """Returns a list of formatted list of commands, including code for tmp storage.
+        """
+        newline = '\n'
+        concat_package_list = ' '.join([repr(str(package)) for package in self.packages_to_install])
+        install_python_packages_script = (
+            f'''if ! [ -x "$(command -v pip)" ]; then{newline}'''
+            f'''    python3 -m ensurepip || python3 -m ensurepip --user || apt-get install python3-pip{newline}'''
+            f'''fi{newline}'''
+            f'''PIP_DISABLE_PIP_VERSION_CHECK=1 python3 -m pip install --quiet \{newline}'''
+            f'''    --no-warn-script-location {concat_package_list} && "$0" "$@"{newline}'''
+            f'''{newline}''')
+        return ['sh', '-c', install_python_packages_script, self.src_code]
+
+    def _create_component_spec(self):
+        """Creates a tmp component scaffold which will be used by the formalize function.
+        Code is temporarily stored in component_spec['implementation']['container']['command'].
+
+        Returns:
+            _type_: _description_ #TODO: FILL OUT
+        """
+        # Instantiate component yaml attributes
+        component_spec = {}
+
+        # Save component name, description, outputs, and parameters
+        component_spec['name'] = self.name
+        if self.description:
+            component_spec['description'] = self.description
+        outputs = self.return_types
+        if outputs:
+            component_spec['outputs'] = outputs
+        component_spec['inputs'] = self.parameters
+
+        # TODO: comment
+        component_spec['implementation'] = {}
+        component_spec['implementation']['container'] = {}
+        component_spec['implementation']['container']['image'] = PLACEHOLDER_IMAGE
+        component_spec['implementation']['container']['command'] = self.packages_to_install_command
+        component_spec['implementation']['container']['args'] = ['--executor_input',
+                                                                {'executorInput': None},
+                                                                '--function_to_execute', 
+                                                                self.name]
+        return component_spec
+
+    def _update_params(self, params: list) -> list:
+        """Converts the parameter types from Python types
+        to Kubeflow types. Currently only supports
+        Python primitive types.
+
+        Args:
+            params: Pipeline parameters. A list of dictionaries,
+                each param is a dict containing keys:
+                    'name': required, str param name.
+                    'type': required, python primitive type.
+                    'description': optional, str param desc.
+        Returns:
+            list: Params list with converted types.
+        Raises:
+            Exception: If an inputted type is not a primitive.
+        """
+        python_kfp_types_mapper = {
+            int: 'Integer',
+            str: 'String',
+            float: 'Float',
+            bool: 'Bool',
+            list: 'JsonArray',
+            dict: 'JsonObject'
+        }
+        for param in params:
+            try:
+                param['type'] = python_kfp_types_mapper[param['type']]
+            except KeyError as err:
+                raise ValueError(f'Unsupported python type - we only support '
+                                f'primitive types at this time. {err}') from err
+        return params
+
+
+class KFPPipeline(BasePipeline):
     """Creates a KFP specific Pipeline object for #TODO: add more
 
     Args:
@@ -135,7 +307,6 @@ class KFPPipeline(Pipeline):
         defaults = read_yaml_file(GENERATED_DEFAULTS_FILE)
         self.project_id = defaults['gcp']['project_id']
         self.gs_pipeline_job_spec_path = defaults['pipelines']['gs_pipeline_job_spec_path']
-
 
         # Build necessary folders
         make_dirs([
@@ -360,3 +531,62 @@ class KFPPipeline(Pipeline):
         # Stringify and sort
         reqs_str = ''.join(r+'\n' for r in sorted(set_of_requirements))
         return reqs_str
+
+
+class KFPServices(BaseServices):
+    """Creates a KFP specific Services object for #TODO: add more
+
+    Args:
+        Services (object): Generic Services object.
+    """
+
+    def __init__(self) -> None:
+        """Initializes KFPServices Object.
+        """
+
+    def _build_dockerfile(self):
+        """Writes the services/submission_service/Dockerfile #TODO add more
+        """
+        # Read in defaults params
+        defaults = read_yaml_file(GENERATED_DEFAULTS_FILE)
+        self.pipeline_storage_path = defaults['pipelines']['pipeline_storage_path']
+        self.pipeline_job_runner_service_account = defaults['gcp']['pipeline_job_runner_service_account']
+        self.pipeline_job_submission_service_type = defaults['gcp']['pipeline_job_submission_service_type']
+        self.project_id = defaults['gcp']['project_id']
+        self.pipeline_job_submission_service_type = defaults['gcp']['pipeline_job_submission_service_type']
+
+        # Set directory for files to be written to
+        self.submission_service_base_dir = BASE_DIR + 'services/submission_service'
+
+        write_file(
+            f'{self.submission_service_base_dir}/Dockerfile', 
+            render_jinja(
+                template_path=import_files(KFP_TEMPLATES_PATH + '.services.submission_service') / 'Dockerfile.j2',
+                base_dir=BASE_DIR,
+                generated_license=GENERATED_LICENSE),
+            'w')
+
+    def _build_requirements(self):
+        """Writes the services/submission_service/requirements.txt #TODO add more
+        """
+        write_file(
+            f'{self.submission_service_base_dir}/requirements.txt', 
+            render_jinja(
+                template_path=import_files(KFP_TEMPLATES_PATH + '.services.submission_service') / 'requirements.txt.j2',
+                pinned_kfp_version=PINNED_KFP_VERSION,
+                pipeline_job_submission_service_type=self.pipeline_job_submission_service_type),
+            'w')
+
+    def _build_main(self):
+        """Writes the services/submission_service/main.py file to #TODO add more
+        """
+        write_file(
+            f'{self.submission_service_base_dir}/main.py', 
+            render_jinja(
+                template_path=import_files(KFP_TEMPLATES_PATH + '.services.submission_service') / 'main.py.j2',
+                generated_license=GENERATED_LICENSE,
+                pipeline_root=self.pipeline_storage_path,
+                pipeline_job_runner_service_account=self.pipeline_job_runner_service_account,
+                pipeline_job_submission_service_type=self.pipeline_job_submission_service_type,
+                project_id=self.project_id),
+            'w')
