+++ google_cloud_automlops/orchestration/templates/kfp/components/component_base/Dockerfile.j2
