+++ google_cloud_automlops/orchestration/templates/kfp/components/component_base/src/task.py.j2
