+++ google_cloud_automlops/orchestration/templates/kfp/pipelines/pipeline.py.j2
