+++ google_cloud_automlops/orchestration/templates/kfp/pipelines/pipeline_runner.py.j2
