+++ google_cloud_automlops/orchestration/templates/kfp/scripts/build_components.sh.j2
