+++ google_cloud_automlops/orchestration/templates/kfp/scripts/build_pipeline_spec.sh.j2
