+++ google_cloud_automlops/orchestration/templates/kfp/scripts/publish_to_topic.sh.j2
