+++ google_cloud_automlops/orchestration/templates/kfp/scripts/run_all.sh.j2
