+++ google_cloud_automlops/orchestration/templates/kfp/scripts/run_pipeline.sh.j2
