+++ google_cloud_automlops/orchestration/templates/kfp/services/submission_service/Dockerfile.j2
