+++ google_cloud_automlops/orchestration/templates/kfp/services/submission_service/main.py.j2
