+++ google_cloud_automlops/utils/constants.py
@@ -121,7 +121,7 @@ PINNED_KFP_VERSION = 'kfp<2.0.0'
 TERRAFORM_TEMPLATES_PATH = 'google_cloud_automlops.provisioning.terraform.templates'
 PULUMI_TEMPLATES_PATH = 'google_cloud_automlops.provisioning.pulumi.templates'
 GCLOUD_TEMPLATES_PATH = 'google_cloud_automlops.provisioning.gcloud.templates'
-KFP_TEMPLATES_PATH = 'google_cloud_automlops.orchestration.kfp.templates'
+KFP_TEMPLATES_PATH = 'google_cloud_automlops.orchestration.templates.kfp'
 CLOUDBUILD_TEMPLATES_PATH = 'google_cloud_automlops.deployments.cloudbuild.templates'
 GITHUB_ACTIONS_TEMPLATES_PATH = 'google_cloud_automlops.deployments.github_actions.templates'
 GITOPS_TEMPLATES_PATH = 'google_cloud_automlops.deployments.gitops.templates'
